add grid iterator

+++ aoc/src/y2019/d08/mod.rs
@@ -39,7 +39,7 @@ fn find_b(v: &[char], size: (usize, usize)) -> String {
                 .map(|layer| layer[y * w + x])
                 .find(|&ch| ch != '2')
                 .unwrap_or('2');
-            grid.set(y, x, if px == '0' { ' ' } else { px });
+            grid.set(y as i64, x as i64, if px == '0' { ' ' } else { px });
         }
     }
     grid.to_string()
